fixed test cases and removed duplicate code
